Feature/add pgsql store

+++ migration/000001_init.db.up.sql
@@ -1,11 +1,11 @@
 CREATE TABLE "users" (
-    "id" SERIAL PRIMARY KEY NOT NULL,
+    "id" BIGSERIAL PRIMARY KEY NOT NULL,
     "telegram_id" bigint UNIQUE,
     "name" VARCHAR (200) NULL
 
 );
 CREATE TABLE "tasks"(
     "id" SERIAL PRIMARY KEY NOT NULL,
-    user_id integer REFERENCES users ON DELETE CASCADE,
+    user_id bigint REFERENCES users ON DELETE CASCADE,
     "text" VARCHAR(255)
 );
